Move utils back to main file
